update: Renew directory structure.
Create `gen` subdirectory.
