+++ src/gen/problem.rs
@@ -1,8 +1,8 @@
 extern crate reqwest;
 extern crate scraper;
 
-use crate::file_utils;
-use crate::sample_cases::SampleCases;
+use crate::gen::file_utils;
+use crate::gen::sample_cases::SampleCases;
 use reqwest::Url;
 use scraper::Html;
 use std::io;
