+++ src/gen/sample_cases.rs
