+++ src/main.rs
@@ -1,31 +1,11 @@
 #[macro_use]
 extern crate clap;
 
-mod problem;
-mod file_utils;
-mod sample_cases;
+mod gen;
 
 use clap::{App, Arg, SubCommand};
 use std::io;
-use problem::Problem;
 
-
-fn execute_fetching_problem(contest_id: &str, problem_id: &str) -> Result<(), io::Error> {
-    let problem = Problem::new(contest_id, &problem_id);
-    if let Some(problem) = problem {
-        problem.create_sample_cases_files()?;
-    }
-    Ok(())
-}
-
-fn execute_fetching_problems_in_contest(contest_id: &str) -> Result<(), io::Error> {
-    let alphabets = (b'a'..=b'z').map(|c| c as char).collect::<Vec<char>>();
-    for alphabet in alphabets {
-        let problem_id = format!("{}", alphabet);
-        execute_fetching_problem(contest_id, &problem_id)?;
-    }
-    Ok(())
-}
 
 fn main() -> Result<(), io::Error> {
     let matches = App::new(crate_name!())
@@ -58,10 +38,10 @@ fn main() -> Result<(), io::Error> {
             // Problem is specified (such as "a", "b", "c"...).
             match (contest_id, problem_id) {
                 (Some(contest_id), Some(problem_id)) => {
-                    execute_fetching_problem(contest_id, &problem_id)?;
+                    gen::execute_fetching_problem(contest_id, &problem_id)?;
                 }
                 (Some(contest_id), None) => {
-                    execute_fetching_problems_in_contest(contest_id)?;
+                    gen::execute_fetching_problems_in_contest(contest_id)?;
                 }
                 (_, _) => {}
             }
